chore: split renderer into files
